Move HttpContent extension methods into HttpClientExtensions
